backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "aiofiles>=24.1.0",
    "motor>=3.7.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]
